chore: add tests, refactor code a bit to make it a tad more testable.
