+++ .github/workflows/build.yml
@@ -15,7 +15,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        node-version: [ 16.x, 18.x, 20.x ]
+        node-version: [ 18.x, 20.x, 22.x ]
         os: [ windows-latest, ubuntu-latest, macOS-latest ]
 
     # Go
@@ -41,7 +41,6 @@ jobs:
       run: npm install
 
     - name: Test
-      if: matrix.node-version > '16.x'
       run: npm test
       env:
         CI: true
