+++ .gitignore
@@ -1,3 +1,5 @@
 scratch/
 node_modules
 .vscode
+.nyc_output
+coverage
