+++ jsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "module": "commonJS",
+    "moduleResolution": "node",
+    "target": "es6"
+  },
+  "include": ["src/**/*"]
+}
